+++ src/main/java/com/bejker/interactionmanager/search/SearchUtil.java
@@ -2,6 +2,8 @@ package com.bejker.interactionmanager.search;
 
 import net.minecraft.block.Block;
 import net.minecraft.client.MinecraftClient;
+import net.minecraft.entity.Entity;
+import net.minecraft.entity.EntityType;
 import net.minecraft.registry.Registries;
 import net.minecraft.registry.entry.RegistryEntry;
 
@@ -11,6 +13,8 @@ import java.util.*;
 public class SearchUtil {
 
     private static SearchTree<Block> blockSearchTree;
+
+    private static SearchTree<EntityType<?>> entitySearchTree;
 
     private static String current_language;
 
@@ -22,23 +26,45 @@ public class SearchUtil {
         }
         current_language = language;
         blockSearchTree = new SearchTree<>();
+        entitySearchTree = new SearchTree<>();
 
         for(var block : Registries.BLOCK){
             RegistryEntry<Block> entry = Registries.BLOCK.getEntry(block);
             blockSearchTree.put(getLocalizedBlockName(block),block);
             blockSearchTree.put(entry.getIdAsString(),block);
         }
+
+        for(var entity_type : Registries.ENTITY_TYPE){
+            RegistryEntry<EntityType<?>> entry = Registries.ENTITY_TYPE.getEntry(entity_type);
+            entitySearchTree.put(getLocalizedEntityName(entity_type),entity_type);
+            entitySearchTree.put(entry.getIdAsString(),entity_type);
+        }
     }
 
     public static Collection<Block> searchBlocks(String word){
         return searchBlocks(word,-1);
     }
+
     public static Collection<Block> searchBlocks(String word,int results){
         init();
         return blockSearchTree.search(word,results);
     }
+
+    public static Collection<EntityType<?>> searchEntities(String word){
+        return searchEntities(word,-1);
+    }
+
+    public static Collection<EntityType<?>> searchEntities(String word,int results){
+        init();
+        return entitySearchTree.search(word,results);
+    }
+
     public static String getLocalizedBlockName(Block block){
         return block.getName().getContent().visit(Optional::of).get().toLowerCase(Locale.ROOT);
+    }
+
+    public static String getLocalizedEntityName(EntityType<?> entityType){
+        return entityType.getName().getContent().visit(Optional::of).get().toLowerCase(Locale.ROOT);
     }
 
     private static class SearchTree<T>{
@@ -75,7 +101,7 @@ public class SearchUtil {
         }
 
         public Collection<T> search(String word,int results){
-            return tree.search(word,results).stream().map(this::mapIndexToEntry).sorted(Comparator.comparing(Object::toString).reversed()).toList();
+            return tree.search(word,results).stream().map(this::mapIndexToEntry).toList();
         }
     }